modelForEventController fixed spring security varsion 1.0.jar is ready for deployment! Alhamdolillah V1.0. Bug fixes. Custom Login form and logout is readye exception handling for 403, 404, 500 resource init Soap notif service init soap working without unmarshalling app now is in a pretty Good shape. Deployment: V1.0 Otp init. im still confused Ui progress: 70% Gorgeous Clock available + backend issues addressed SOME FIXES AND UX IMPROVEMENT OTP Working Just fine This is version 2.0 - Deployed version 2.1 init - Revolution in task and centre modules center package almost completed. Awaiting test TaskServiceImpl modifying is in progress TaskServiceImpl is almost ready. Last part of this revolution: views and relates methods Some changes on Person (Experimental) - Utils has more useful methods
Etisalat init

DB config and related changes in code

Temperature submit available

DASHBOARD contd 2
File is ready for event

Pm - center modification init

persistence init

Pm controller and service contd plus persistence model verification
Models are almost ready. few consideration about persistence
Model created with no error. temperature, event with resource and daily report pending

+++ src/main/java/ir/tic/clouddc/pm/TemperaturePm.java
@@ -83,6 +83,7 @@ public class TemperaturePm extends Pm {
         pm.setFinishedDate(UtilService.getDATE());
         pm.setFinishedTime(UtilService.getTime());
         pm.setActive(false);
+
         TemperaturePm temperaturePm = (TemperaturePm) pm;
         List<TemperaturePmDetail> temperaturePmDetailList = new ArrayList<>();
         for (PmDetail pmDetail : temperaturePm.getPmDetailList()) {
@@ -90,7 +91,7 @@ public class TemperaturePm extends Pm {
         }
         float sum = (float) temperaturePmDetailList.stream().filter(temperaturePmDetail -> temperaturePmDetail.getTemperatureValue() > 0.0f).mapToDouble(TemperaturePmDetail::getTemperatureValue).sum();
         int reportedTemperatures = (int) temperaturePmDetailList.stream().filter(temperaturePmDetail -> temperaturePmDetail.getTemperatureValue() > 0.0f).count();
-        temperaturePm.setAverageDailyValue(sum / reportedTemperatures);
+        temperaturePm.setAverageDailyValue(sum / (float) reportedTemperatures);
 
         return temperaturePm;
     }
